refactor: Upgrading to Laravel Vite and code splitting
