+++ package.json
@@ -1,18 +1,17 @@
 {
     "private": true,
     "scripts": {
-        "tailwind-prod": "npx tailwindcss -i resources/css/app.css -o public/css/app.css --minify",
         "dev": "vite",
         "build": "vite build"
     },
     "devDependencies": {
         "autoprefixer": "^10.2.6",
         "axios": "^0.21",
+        "laravel-vite-plugin": "^0.6.0",
         "lodash": "^4.17.19",
         "postcss": "^8.3.5",
         "tailwindcss": "^2.2.4",
-        "vite": "^3.0.2",
-        "laravel-vite-plugin": "^0.6.0"
+        "vite": "^3.0.2"
     },
     "dependencies": {
         "@tailwindcss/forms": "^0.3.3",
